feat: table refactoring
